+++ samples/api.pipe.js
@@ -3,11 +3,11 @@ const parse = require('..')
 const generate = require('csv-generate')
 const transform = require('stream-transform')
 
+const generator = generate({
+  length: 20
+})
 const parser = parse({
   delimiter: ':'
-})
-const input = generate({
-  length: 20
 })
 const transformer = transform(function(record, callback){
   setTimeout(function(){